backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
openai>=1.12.0
langchain>=0.1.0
langgraph>=0.0.24
chromadb>=0.4.22
neo4j>=5.17.0
python-dotenv>=1.0.1
requests>=2.31.0
pytest>=8.0.0
google-generativeai>=0.3.2
